Initial commit
Doesn’t do anything yet, but here is the general setup.

The idea here it to implement the MVC design pattern by separating Game
and its controller and view classes.

The next step is to implement the CardLayout so we can load the
necessary view panels depending on the game state without having to
show a new window.