Place Plants | Upgrade Button | Drawing bugs
The user can now place plants on the estuary. When the mouse moves, the
radius will be displayed.

The upgrade button at the top can be pressed to skip straight to the
next upgrade stage. Make sure the game starts first

Fixed small bugs with estuary objects so they are now positioned on the
center of the object